# Fixed .shallow to consistently retain keys and indices.