--- a/NEWS.md
+++ b/NEWS.md
@@ -78,6 +78,9 @@
 
 17. `.SD` would incorrectly include symbol on lhs of `:=` when `.SDcols` is specified and `get()` appears in `j`. Thanks @renkun-ken for reporting and the PR. Closes #2326.
 
+18. .shallow retains keys correctly [#2336](https://github.com/Rdatatable/data.table/issues/2336)
+    Thanks to @MarkusBonsch for reporting and fixing
+    
 #### NOTES
 
 1. `?data.table` makes explicit the option of using a `logical` vector in `j` to select columns, [#1978](https://github.com/Rdatatable/data.table/issues/1978). Thanks @Henrik-P for the note and @MichaelChirico for filing.