--- a/inst/tests/tests.Rraw
+++ b/inst/tests/tests.Rraw
@@ -6846,6 +6846,7 @@ test(1544.2, setDF(merge(x1, y, by="a2")), merge(as.data.frame(x1), as.data.fram
 x1 <- data.table(a1 = c('a', 'b', 'c'), a2 = c(1L, 3L, 2L), a3 = c(TRUE, FALSE, TRUE), key="a1,a2")
 test(1545.1, key(.shallow(x1, cols="a2")), NULL)
 test(1545.2, key(.shallow(x1, retain.key=FALSE)), NULL)
+test(1545.2, key(.shallow(x1, cols = "a1", retain.key=FALSE)), NULL)
 test(1545.3, key(.shallow(x1, retain.key=TRUE)), key(x1))
 test(1545.4, key(.shallow(x1, cols="a1", retain.key=TRUE)), "a1")
 # tests for #2336. .shallow drops keys unnecessarily
@@ -6857,7 +6858,9 @@ test(1545.9, key(.shallow(x1, cols=c("a2"), retain.key=TRUE)), NULL)
 test(1545.10, key(.shallow(x1, cols=c("a3"), retain.key=TRUE)), NULL)
 setkey(x1, NULL)
 test(1545.11, key(.shallow(x1, retain.key=TRUE)), NULL)
+test(1545.111, key(.shallow(x1, retain.key=FALSE)), NULL)
 test(1545.12, key(.shallow(x1, cols=c("a1", "a2"), retain.key=TRUE)), NULL)
+test(1545.121, key(.shallow(x1, cols=c("a1", "a2"), retain.key=FALSE)), NULL)
 x1 <- x1[0]
 test(1545.13,  key(.shallow(x1, retain.key=TRUE)), NULL)
 test(1545.131, key(.shallow(x1, retain.key=FALSE)), NULL)
@@ -6869,9 +6872,7 @@ test(1545.135, key(.shallow(x1, cols = "a2", retain.key=FALSE)), NULL)
 test(1545.136, key(.shallow(x1, retain.key=TRUE)), "a1")
 test(1545.137, key(.shallow(x1, cols=c("a1", "a2"), retain.key=TRUE)), "a1")
 test(1545.138, key(.shallow(x1, cols=c("a3"), retain.key=TRUE)), NULL)
-setkey(x1, NULL)
 
-test(1545.136, key(.shallow(x1, retain.key=FALSE)), NULL)
 # tests for #2336. .shallow now retains indices as well
 x1 <- data.table(a1 = c('a', 'a', 'a', 'a', 'b', 'c'), a2 = c(2L, 2L, 1L, 2L, 3L, 2L), a3 = c(FALSE, TRUE, TRUE, FALSE, FALSE, TRUE), key="a1,a2")
 setindex(x1, a1, a2, a3)