@MarkusBonsch 2017-09-08
Gets a problem on foverlaps test in travis that I don't get locally. Will need to investigate.