@MarkusBonsch 2017-09-09
Initially, my local tests didn't include fOverlaps since genomicRanges was not installed. Now, everything should be fixed.